refactor: reorganize component imports and update paths for 2024 structure
