+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import ThankYou2024 from "@/components/thank-you-2024";
+import ThankYou2024 from "@/components/2024/thank-you-2024";
 
 export default function Home() {
   return <ThankYou2024 />;
