+++ src/components/2024/EventSchedule/PrivateNetwork.tsx
