+++ src/components/2024/dp/crop-image.tsx
