+++ src/components/2024/home/community-partners/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import SectionTitle from "@/components/section-title";
-import { InfiniteMovingCards } from "@/components/infinite-moving-cards";
-import { communityPartners } from "@/data";
+import SectionTitle from "@/components/2024/section-title";
+import { InfiniteMovingCards } from "@/components/2024/infinite-moving-cards";
+import { communityPartners } from "@/data/2024";
 
 const CommunityPartners = () => {
   return (
