+++ src/components/2024/home/home-hero/index.tsx
@@ -2,8 +2,8 @@ import { Button } from "@chakra-ui/react";
 import Link from "next/link";
 import React from "react";
 import { BsArrowUpRight } from "react-icons/bs";
-import ConfettiContainer from "@/components/confetti/ConfettiContainer";
-import { ctaLinks } from "@/data";
+import ConfettiContainer from "@/components/2024/confetti/ConfettiContainer";
+import { ctaLinks } from "@/data/2024/global";
 // import Image from "next/image";
 // import sunset from "@public/icons/icon-sunset.svg";
