+++ src/components/2024/speakers-card/index.tsx
