+++ src/data/2024/partners.ts
