+++ src/data/2024/speakers.ts
